+++ tests/service/stress/units/testDeployContract.go
@@ -2,15 +2,16 @@ package units
 
 import (
 	"context"
-
 	"encoding/json"
 
 	"github.com/consensys/orchestrate/pkg/errors"
 	orchestrateclient "github.com/consensys/orchestrate/pkg/sdk/client"
+	clientutils "github.com/consensys/orchestrate/pkg/toolkit/app/http/client-utils"
 	"github.com/consensys/orchestrate/pkg/toolkit/app/log"
 	"github.com/consensys/orchestrate/pkg/types/api"
 	"github.com/consensys/orchestrate/pkg/types/tx"
 	"github.com/consensys/orchestrate/pkg/utils"
+	"github.com/consensys/orchestrate/services/api/service/controllers"
 	utils2 "github.com/consensys/orchestrate/tests/service/stress/utils"
 	utils3 "github.com/consensys/orchestrate/tests/utils"
 	"github.com/consensys/orchestrate/tests/utils/chanregistry"
@@ -39,7 +40,9 @@ func BatchDeployContractTest(ctx context.Context, cfg *WorkloadConfig, client or
 	sReq, _ := json.Marshal(req)
 
 	logger = logger.WithField("chain", req.ChainName).WithField("idem", idempotency)
-	_, err := client.SendDeployTransaction(ctx, req)
+	_, err := client.SendDeployTransaction(context.WithValue(ctx, clientutils.RequestHeaderKey, map[string]string{
+		controllers.IdempotencyKeyHeader: idempotency,
+	}), req)
 
 	if err != nil {
 		if !errors.IsConnectionError(err) {